Refactor the GoogleSearcherClass. Extract dom manipulation methods in a trait.
